+++ spec/rubocop/cop/workit/committee_expected_response_status_code_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
+RSpec.describe RuboCop::Cop::Workit::CommitteeExpectedResponseStatusCode, :config do
   it "registers an offense and autocorrects when using `assert_schema_conform` " \
      "with no argument and `have_http_status`" do
     expect_offense(<<~RUBY)
@@ -9,13 +9,14 @@ RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
         expect(response).to have_http_status 400
         do_something
         assert_schema_conform
-        ^^^^^^^^^^^^^^^^^^^^^ Pass expected response status code to check it against the corresponding schema explicitly.
+        ^^^^^^^^^^^^^^^^^^^^^ Specify the HTTP status code of the expected response as an argument.
       end
     RUBY
 
     expect_correction(<<~RUBY)
       it 'something' do
         subject
+        expect(response).to have_http_status 400
         do_something
         assert_schema_conform(400)
       end
@@ -29,7 +30,7 @@ RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
         subject
         do_something
         assert_schema_conform
-        ^^^^^^^^^^^^^^^^^^^^^ Pass expected response status code to check it against the corresponding schema explicitly.
+        ^^^^^^^^^^^^^^^^^^^^^ Specify the HTTP status code of the expected response as an argument.
       end
     RUBY
 
@@ -54,13 +55,14 @@ RSpec.describe RuboCop::Cop::Workit::ComitteeAssertSchemaConfirm, :config do
         expect(response).to have_http_status 400
         do_something
         assert_response_schema_confirm
-        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Pass expected response status code to check it against the corresponding schema explicitly.
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Specify the HTTP status code of the expected response as an argument.
       end
     RUBY
 
     expect_correction(<<~RUBY)
       it 'something' do
         subject
+        expect(response).to have_http_status 400
         do_something
         assert_response_schema_confirm(400)
       end